requirements.txt
dnspython==2.1.0
numpy==1.19.5
pandas==1.1.5
pymongo==3.12.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
